+++ slider/platformio.ini
@@ -18,9 +18,18 @@ lib_deps =
 	marcoschwartz/LiquidCrystal_I2C@^1.1.4
 	waspinator/AccelStepper@^1.64
 
-[env:simulator]
+[env:simulator-receiver]
 build_flags = 
 	-D IS_SIMULATOR=TRUE
+	-D IS_RECEIVER=TRUE
+	-I ../slider
+debug_build_flags = -O0 -g -ggdb
+build_type = debug
+
+[env:simulator-sender]
+build_flags = 
+	-D IS_SIMULATOR=TRUE
+	-D IS_SENDER=TRUE
 	-I ../slider
 debug_build_flags = -O0 -g -ggdb
 build_type = debug
@@ -30,7 +39,9 @@ upload_port = COM3
 monitor_port = COM3
 upload_speed = 921600
 board_build.flash_mode = qio
-build_flags = -I ../slider
+build_flags = 
+	-D IS_RECEIVER=TRUE
+	-I ../slider
 build_type = release
 
 [env:sender]
@@ -38,5 +49,7 @@ upload_port = COM4
 monitor_port = COM4
 upload_speed = 921600
 board_build.flash_mode = qio
-build_flags = -I ../slider
+build_flags = 
+	-D IS_SENDER=TRUE
+	-I ../slider
 build_type = release
