+++ slider/src/test/network/netApp.h
@@ -4,6 +4,17 @@
 #include "src/app/appBase.h"
 #include "src/app/appConfig.h"
 
+struct HandshakeMessage
+{
+    uint8_t mac[6];
+};
+
+struct InputMessage
+{
+    int x;
+    int y;
+};
+
 class NetApp : public Slider::AppBase
 {
     public:
@@ -13,6 +24,9 @@ class NetApp : public Slider::AppBase
 
         void UpdateSender();
         void UpdateReceiver();
+        
+        void Sender__ReceiveMsg(HandshakeMessage message);
+        void Sender__ReceiveMsg(InputMessage message);
 
     private:
         bool isReceiver;
